Added confirm to claiming into organization
