+++ frontend/src/buttons/ConfirmButton.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState, forwardRef } from 'react'
 import { IconButton, ButtonProps } from './IconButton'
 import { Confirm, ConfirmProps } from '../components/Confirm'
 
@@ -7,22 +7,22 @@ type Props = ButtonProps & {
   confirmProps?: Omit<ConfirmProps, 'open' | 'onConfirm' | 'onDeny'>
 }
 
-export const ConfirmButton: React.FC<Props> = ({ onClick, confirm, confirmProps, ...props }) => {
-  const [open, setOpen] = React.useState<boolean>(false)
+export const ConfirmButton = forwardRef<HTMLButtonElement, Props>(({ onClick, confirm, confirmProps, ...props }, ref) => {
+  const [open, setOpen] = useState<boolean>(false)
 
-  const handleClick = e => {
+  const handleClick = (e: React.MouseEvent) => {
     if (confirm) setOpen(true)
     else onClick?.(e)
   }
 
-  const handleConfirm = e => {
+  const handleConfirm = (e: React.MouseEvent) => {
     onClick?.(e)
     setOpen(false)
   }
 
   return (
     <>
-      <IconButton {...props} onClick={handleClick} />
+      <IconButton ref={ref} {...props} onClick={handleClick} />
       {confirm && onClick && (
         <Confirm {...confirmProps} open={open} onConfirm={handleConfirm} onDeny={() => setOpen(false)}>
           {confirmProps?.children}
@@ -30,4 +30,4 @@ export const ConfirmButton: React.FC<Props> = ({ onClick, confirm, confirmProps,
       )}
     </>
   )
-}
+})
